requirements.txt
numpy
pandas
scipy
scikit-learn
matplotlib
pyyaml
pytest
fastapi
uvicorn
jupyter
notebook
joblib
requests
